// src/blockstore/types.ts
export type AnyLink = Link<unknown, number, number, Version>;
export type CarGroup = AnyLink[];
export type CarLog = CarGroup[];
export type AnyAnyLink = Link<unknown, number, number, Version>;

export type AnyLinkFn = (cid: AnyLink) => Promise<AnyBlock | undefined>;

export interface AnyBlock {
  readonly cid: Link<unknown, number, number, Version>;
  readonly bytes: Uint8Array;
}
// ...
export interface IvAndBytes {
  readonly bytes: Uint8Array;
  readonly iv: Uint8Array;
}

export interface BytesWithIv {
  readonly bytes: Uint8Array;
  readonly iv?: Uint8Array;
}
// ...
export type TransactionMeta = unknown;
// ...
export interface KeyedCrypto {
  readonly logger: Logger;
  readonly crypto: CryptoRuntime;
  // readonly codec: BlockCodec<number, IvAndBytes>;
  readonly isEncrypting: boolean;
  fingerPrint(): Promise<string>;
  algo(iv?: Uint8Array): { name: string; iv: Uint8Array; tagLength: number };
  codec(iv?: Uint8Array): BlockCodec<number, Uint8Array>;
  _decrypt(data: IvAndBytes): Promise<Uint8Array>;
  _encrypt(data: BytesWithIv): Promise<Uint8Array>;
  // encode(data: Uint8Array): Promise<Uint8Array>;
  // decode(bytes: Uint8Array | ArrayBuffer): Promise<Uint8Array>;
}
// ...
export interface StoreOpts extends StoreFactory {
  readonly isIndex?: string; // index prefix
  readonly stores?: {
    // string means local storage
    // URL means schema selects the storeType
    readonly useEncryptedBlockstore?: boolean;
    readonly base?: CoerceURI;

    readonly meta?: CoerceURI;
    readonly data?: CoerceURI;
    readonly index?: CoerceURI;
    readonly wal?: CoerceURI;
  };
}

export interface StoreRuntime {
  // the factories should produce ready-to-use stores
  // which means they have to call start() on the store
  // to fullfill lifecycle requirements
  // to release resources, like one database connection
  // for all stores a refcount on close() should be used
  makeMetaStore(loader: Loadable): Promise<MetaStore>;
  makeDataStore(loader: Loadable): Promise<DataStore>;
  makeWALStore(loader: Loadable): Promise<WALStore>;
  encodeFile(blob: BlobLike): Promise<{ cid: AnyLink; blocks: AnyBlock[] }>;
  decodeFile(blocks: unknown, cid: AnyLink, meta: DocFileMeta): Promise<File>;
}

export interface CommitOpts {
  readonly noLoader?: boolean;
  readonly compact?: boolean;
  // readonly public?: boolean;
}

export interface DbMeta {
  readonly cars: CarGroup;
  // key?: string;
}
// ...
export type LoadHandler = (dbMetas: DbMeta[]) => Promise<void>;
// ...
export interface BaseStore {
  readonly storeType: StoreType;
  // readonly url: URI
  url(): URI;
  readonly name: string;
  onStarted(fn: () => void): void;
  onClosed(fn: () => void): void;

  keyedCrypto(): Promise<KeyedCrypto>;

  close(): Promise<Result<void>>;
  destroy(): Promise<Result<void>>;
  readonly ready?: () => Promise<void>;
  start(): Promise<Result<URI>>;
}

export interface MetaStore extends BaseStore {
  readonly storeType: "meta";
  load(branch?: string): Promise<DbMeta[] | Falsy>;
  // branch is defaulted to "main"
  save(meta: DbMeta, branch?: string): Promise<Result<void>>;
}

export interface RemoteMetaStore extends MetaStore {
  handleByteHeads(byteHeads: Uint8Array[], branch?: string): Promise<DbMeta[]>;
}

export interface DataSaveOpts {
  readonly public: boolean;
}

export interface DataStore extends BaseStore {
  readonly storeType: "data";
  load(cid: AnyLink): Promise<AnyBlock>;
  save(car: AnyBlock, opts?: DataSaveOpts): Promise</*AnyLink | */ void>;
  remove(cid: AnyLink): Promise<Result<void>>;
}

export interface WALState {
  operations: DbMeta[];
  noLoaderOps: DbMeta[];
  fileOperations: {
    readonly cid: AnyLink;
    readonly public: boolean;
  }[];
}

export interface WALStore extends BaseStore {
  readonly storeType: "wal";
  ready: () => Promise<void>;
  readonly processing?: Promise<void> | undefined;
  readonly processQueue: CommitQueue<void>;

  process(): Promise<void>;
  enqueue(dbMeta: DbMeta, opts: CommitOpts): Promise<void>;
  enqueueFile(fileCid: AnyLink /*, publicFile?: boolean*/): Promise<void>;
  load(): Promise<WALState | Falsy>;
  save(state: WALState): Promise<void>;
}

export type CompactFetcher = BlockFetcher & {
  readonly loggedBlocks: CarTransaction;
};
export type CompactFn = (blocks: CompactFetcher) => Promise<TransactionMeta>;
// ...
export interface BlockstoreRuntime {
  readonly logger: Logger;
  readonly applyMeta: (meta: TransactionMeta, snap?: boolean) => Promise<void>;
  readonly compact: CompactFn;
  readonly autoCompact: number;
  readonly crypto: CryptoRuntime;
  readonly store: StoreOpts;
  readonly storeRuntime: StoreRuntime;
  readonly keyBag: Partial<KeyBagOpts>;
  // readonly public: boolean;
  readonly meta?: DbMeta;
  readonly name?: string;
  readonly threshold: number;
}

export interface Loadable {
  readonly name: string; // = "";
  readonly logger: Logger;
  readonly ebOpts: BlockstoreRuntime;
  remoteCarStore?: DataStore;
  carStore(): Promise<DataStore>;
  carLog: CarLog; // = new Array<CarGroup>();
  remoteMetaStore?: RemoteMetaStore;
  remoteFileStore?: DataStore;
  ready(): Promise<void>;
  close(): Promise<void>;
  fileStore(): Promise<DataStore>;
  WALStore(): Promise<WALStore>;
  handleDbMetasFromStore(metas: DbMeta[]): Promise<void>;
}